TE-448 Use dynamic web port: Fix bootstrap name and build infra
