+++ rcp/org.testeditor.rcp4.bootstrap/.settings/org.eclipse.jdt.core.prefs
