+++ rcp/org.testeditor.rcp4.bootstrap/build.properties
