+++ rcp/org.testeditor.rcp4.bootstrap/pom.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+		<groupId>org.testeditor</groupId>
+		<artifactId>org.testeditor.releng.parent</artifactId>
+		<version>1.2.0-SNAPSHOT</version>
+		<relativePath>../../releng/org.testeditor.releng.parent</relativePath>
+	</parent>
+
+	<artifactId>org.testeditor.rcp4.bootstrap</artifactId>
+	<packaging>eclipse-plugin</packaging>
+
+	<build>
+		<plugins>
+			<plugin>
+				<groupId>org.eclipse.xtend</groupId>
+				<artifactId>xtend-maven-plugin</artifactId>
+			</plugin>
+		</plugins>
+	</build>
+</project>
+
