+++ rcp/pom.xml
@@ -27,6 +27,7 @@
 		<module>org.testeditor.rcp4.platform.feature</module>
 		<module>org.testeditor.rcp4.updatesite</module>
 		<module>org.testeditor.rcp4.feature</module>
+		<module>org.testeditor.rcp4.bootstrap</module>
 		<module>org.testeditor.rcp4.tcltestrun</module>
 		<module>org.testeditor.rcp4.views.tcltestrun</module>
 		<module>org.testeditor.rcp4.tcltestrun.tests</module>
